ai-agent-server/requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# LLM Clients
anthropic==0.40.0       # Claude API (개발/테스트)
httpx==0.27.2           # Loki/Prometheus/vLLM HTTP 호출

# Aurora PostgreSQL
psycopg2-binary==2.9.10

# Settings
pydantic-settings==2.6.1
